Don't interpolate view when panning
Also make renderer update take a double rather than long since we can potentially have <1ms frames

+++ src/MapEditor/MapEditContext.h
@@ -77,7 +77,7 @@ public:
 	void lockMouse(bool lock);
 
 	// General
-	bool update(long frametime);
+	bool update(double frametime);
 
 	// Map loading
 	bool openMap(const Archive::MapDesc& map);
